configured clock to 425kHZ, needs to be 400k

+++ Distance meter/Drivers/OLED/ssd1306.h
@@ -209,7 +209,7 @@ uint8_t ssd1306_GetDisplayOn();
 // Low-level procedures
 void ssd1306_Reset(void);
 void ssd1306_WriteCommand(uint8_t byte);
-void ssd1306_WriteData(uint8_t* buffer, size_t buff_size);
+void ssd1306_WriteData(uint8_t* buffer, uint16_t buff_size);
 SSD1306_Error_t ssd1306_FillBuffer(uint8_t* buf, uint32_t len);
 
 _END_STD_C
